code refactored - updated get fruit integration test - removed PagedView.cs file as we are not using paging - move files from Sample.Api to Sample.Test which belongs to tests
