+++ Sample.Api/Controllers/FruitController.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Mvc;
 using Sample.Api.Config;
 using Sample.Api.DTO;
