--- Sample.Api/DTO/PagedView.cs
@@ -1,9 +1,0 @@
-using System;
-namespace Sample.Api.DTO
-{
-    public class PagedView<T> where T : class
-    {
-        public List<T> objectList { get; set; }
-    }
-}
-
