+++ Sample.Api/Program.cs
@@ -20,4 +20,5 @@ app.UseMiddleware<HttpLoggingMiddleware>();
 
 app.Run();
 
+// Integration Tests hook
 public partial class Program { }
