+++ Sample.Tests/Sample.Tests.csproj
@@ -31,8 +31,4 @@
       <Folder Include="Unit\" />
       <Folder Include="Integration\" />
     </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="Integration\" />
-    </ItemGroup>
 </Project>
